Commented ctrl+c mapping
Commented ctrl+c mapping because it overrides the real CTRL+C purpose that is used for interrupting blocking processes. In case the mapping in the conf file is enabled, if a process is blocked in the terminal, the user cannot interrupt it and take control of the terminal again. For this reason I suggest to comment/remove those mappings.

The user can copy the content in the terminal by the default keybinding that is `CTRL+SHIFT+C` (same for paste by CTRL+SHIFT+V).