Update spaceleft.js
Bug:
The Hyprland.active.client.title can be too long to attach closely to the next widget (normally when a browser's tab has too long title).

I just change the substring the long title and let it end with "..."

Should be a more graceful way to handle this. But that's beyond my scope.