changed cheatsheet to appear only on active workspace
When loading the cheatsheet on multiple monitors, using the PgUp and PgDn keys to cycle through the pages doesn't behave as expected. The page switch is sent to the different monitors randomly, and I have to spam the key until it happens to flip the page on the active workspace.