Update workspace_action.sh
This script is kinda an improvement for the workspace_action script of end4. This script allows for relative workspace keybind. The next and prev workspaces allow for relative Workspace support, using the second arg. while the first arg is "ch" or "mv"

i made this script as i found it annoying that when using the traditional keybind, bind = ... , workspace, r+1/r-1, hyprland changed 4 workspaces instead of 1. And upon investigation, it seems that using hyprctl is better. but the original workspace_action didnt work for relative workspace. and was frankly quite obfuscated. So this is a better script in my opinion.